backend/requirements.txt
black
flake8
isort
requests
fastapi
uvicorn[standard]
mangum
boto3
pytz
python-jose
